Document `ProjectiveCoordinates`
